+++ dds/notifications.py
@@ -98,7 +98,8 @@ class _DDHNotification:
         self.ddh_bluez_version = g_bluez_ver
         if g:
             lat, lon, _, speed = g
-            self.ddh_gps_position = '{:.4f}, {:.4f}'.format(float(lat), float(lon))
+            if lat:
+                self.ddh_gps_position = '{:.4f}, {:.4f}'.format(float(lat), float(lon))
             self.ddh_gps_speed = '{:.2f} knots'.format(float(speed))
         self.ddh_box_name = dds_get_cfg_vessel_name()
         self.ddh_box_sn = dds_get_cfg_box_sn()