+++ CHANGELOG.md
@@ -65,3 +65,9 @@
     maps GIFs to .gitignore
     pop_ddh2.sh -> pop_ddh.sh
 
+4.0.14
+
+    fixed notification GPS error from Nick
+    https://app.asana.com/0/1206361785539879/1208148157218367/f
+    rearranged a bit main_gui and utils_gui.py
+